Ansible filter mismatch is probably not a problem
It seems reasonable to have no scenarios match when testing a given
role/driver combination in automation. Given the specific exit code to
check for this (101), I think the message that gets printed does not
need to indicate that "this is a problem", so I removed that.

While I was in there, I switched it over to using the builtin tox
terminal reporter instance for logging, so the printed message should
now also respect all tox reporting settings that a user may have.